requirements.txt
streamlit
streamlit-pills
openai

streamlit-aggrid==0.2.2-2

snowflake-connector-python==2.7.7
snowflake-sqlalchemy
sqlalchemy

#For Postgres
psycopg2-binary